@@ -23,15 +23,12 @@
             "role": "Project Founder"
         }
     ],
-    "support": {
-        "forum": "https://groups.google.com/forum/#!forum/twig-users"
-    },
     "require": {
-        "php": ">=5.3.3",
+        "php": ">=5.4.0",
         "symfony/polyfill-ctype": "^1.8"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "^3.3",
+        "symfony/phpunit-bridge": "^3.4.19|^4.1.8",
         "symfony/debug": "^2.7",
         "psr/container": "^1.0"
     },
@@ -45,7 +42,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.35-dev"
+            "dev-master": "1.38-dev"
         }
     }
 }
