@@ -2,6 +2,9 @@
 
 namespace Drupal\twig_extension_test\TwigExtension;
 
+use Twig\TwigFilter;
+use Twig\TwigFunction;
+
 /**
  * A test Twig extension that adds a custom function and a custom filter.
  */
@@ -21,7 +24,7 @@ class TestExtension extends \Twig_Extension {
    */
   public function getFunctions() {
     return [
-      'testfunc' => new \Twig_Function_Function(['Drupal\twig_extension_test\TwigExtension\TestExtension', 'testFunction']),
+      'testfunc' => new TwigFunction('testfunc', ['Drupal\twig_extension_test\TwigExtension\TestExtension', 'testFunction']),
     ];
   }
 
@@ -39,7 +42,7 @@ class TestExtension extends \Twig_Extension {
    */
   public function getFilters() {
     return [
-      'testfilter' => new \Twig_Filter_Function(['Drupal\twig_extension_test\TwigExtension\TestExtension', 'testFilter']),
+      'testfilter' => new TwigFilter('testfilter', ['Drupal\twig_extension_test\TwigExtension\TestExtension', 'testFilter']),
     ];
   }
 
