@@ -26,13 +26,6 @@ class EntityUpdateAddRevisionTranslationAffectedTest extends UpdatePathTestBase 
   protected $entityManager;
 
   /**
-   * The last installed schema repository service.
-   *
-   * @var \Drupal\Core\Entity\EntityLastInstalledSchemaRepositoryInterface
-   */
-  protected $lastInstalledSchemaRepository;
-
-  /**
    * The state service.
    *
    * @var \Drupal\Core\State\StateInterface
@@ -45,8 +38,8 @@ class EntityUpdateAddRevisionTranslationAffectedTest extends UpdatePathTestBase 
   protected function setUp() {
     parent::setUp();
 
+    // Do not use this property after calling ::runUpdates().
     $this->entityManager = \Drupal::entityManager();
-    $this->lastInstalledSchemaRepository = \Drupal::service('entity.last_installed_schema.repository');
     $this->state = \Drupal::state();
   }
 
@@ -71,19 +64,19 @@ class EntityUpdateAddRevisionTranslationAffectedTest extends UpdatePathTestBase 
 
     // Check that the test entity type does not have the
     // 'revision_translation_affected' field before running the updates.
-    $field_storage_definitions = $this->lastInstalledSchemaRepository->getLastInstalledFieldStorageDefinitions('entity_test_update');
+    $field_storage_definitions = \Drupal::service('entity.last_installed_schema.repository')->getLastInstalledFieldStorageDefinitions('entity_test_update');
     $this->assertFalse(isset($field_storage_definitions['revision_translation_affected']));
 
     $this->runUpdates();
 
     // Check that the 'revision_translation_affected' field has been added by
     // system_update_8402().
-    $field_storage_definitions = $this->lastInstalledSchemaRepository->getLastInstalledFieldStorageDefinitions('entity_test_update');
+    $field_storage_definitions = \Drupal::service('entity.last_installed_schema.repository')->getLastInstalledFieldStorageDefinitions('entity_test_update');
     $this->assertTrue(isset($field_storage_definitions['revision_translation_affected']));
 
     // Check that the correct initial value was set when the field was
     // installed.
-    $entity = $this->entityManager->getStorage('entity_test_update')->load(1);
+    $entity = \Drupal::entityTypeManager()->getStorage('entity_test_update')->load(1);
     $this->assertTrue($entity->revision_translation_affected->value);
   }
 
