@@ -110,13 +110,13 @@ class TimestampAgoFormatterTest extends BrowserTestBase {
     $edit = [
       'fields[field_timestamp][settings_edit_form][settings][future_format]' => 'ends in @interval',
       'fields[field_timestamp][settings_edit_form][settings][past_format]' => 'started @interval ago',
-      'fields[field_timestamp][settings_edit_form][settings][granularity]' => 3,
+      'fields[field_timestamp][settings_edit_form][settings][granularity]' => 1,
     ];
     $this->drupalPostForm(NULL, $edit, 'Update');
     $this->drupalPostForm(NULL, [], 'Save');
 
-    $this->assertSession()->pageTextContains('ends in 1 year 1 month 1 week');
-    $this->assertSession()->pageTextContains('started 1 year 1 month 1 week ago');
+    $this->assertSession()->pageTextContains('ends in 1 year');
+    $this->assertSession()->pageTextContains('started 1 year ago');
   }
 
 }
